Build: option added to use much more compiler warnings for GCC, fixed some issues

+++ src/marnav/nmea/date.cpp
@@ -77,7 +77,7 @@ date date::parse(const std::string & str) throw(std::invalid_argument)
 {
 	try {
 		std::size_t pos = 0;
-		uint32_t t = std::stoul(str, &pos);
+		uint32_t t = std::stoi(str, &pos);
 		if (pos != str.size())
 			throw std::invalid_argument{"invalid format for date"};
 		return date{t % 100, static_cast<month>((t / 100) % 100), (t / 10000) % 100};
